lots of refactoring in traits to isolate functional components, vars, and vals
