+++ server/src/main/scala/com/ossuminc/riddl/lsp/server/RiddlLSPServer.scala
@@ -1,30 +1,42 @@
 package com.ossuminc.riddl.lsp.server
 
-import com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService.RiddlLSPTextDocumentService
-import com.ossuminc.riddl.lsp.utils.createInitializeResultIncremental
-import org.eclipse.lsp4j.{CompletionOptions, InitializeParams, InitializeResult, ServerCapabilities, TextDocumentSyncKind}
-import org.eclipse.lsp4j.services.{LanguageServer, TextDocumentService, WorkspaceService}
+import com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService.DocContentMgmt.{
+  ActionsAndLenses,
+  CompletionAndHover
+}
+import com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService.{
+  DocLifecycleMgmt,
+  RiddlLSPTextDocumentService,
+  Vars
+}
+import com.ossuminc.riddl.lsp.utils.Utils
+import com.ossuminc.riddl.lsp.utils.Utils.createInitializeResultIncremental
+import org.eclipse.lsp4j.{CompletionOptions, InitializeParams, InitializeResult}
+import org.eclipse.lsp4j.services.{
+  LanguageServer,
+  TextDocumentService,
+  WorkspaceService
+}
 
 import java.util.concurrent.CompletableFuture
 
 class RiddlLSPServer extends LanguageServer {
-  val documentService = new RiddlLSPTextDocumentService()
-  val workspaceService = new RiddlLSPWorkspaceService()
+  private val documentService = new RiddlLSPTextDocumentService()
+  private val workspaceService = new RiddlLSPWorkspaceService()
 
-  override def initialize(params: InitializeParams): CompletableFuture[InitializeResult] = {
+  override def initialize(
+      params: InitializeParams
+  ): CompletableFuture[InitializeResult] = {
     val result = createInitializeResultIncremental()
     CompletableFuture.supplyAsync(() => result)
   }
 
-  override def shutdown(): CompletableFuture[AnyRef] = {
+  override def shutdown(): CompletableFuture[AnyRef] =
     CompletableFuture.supplyAsync(() => null)
-  }
 
-  override def exit(): Unit = {
-    ()
-  }
+  override def exit(): Unit = ()
 
-  override def getTextDocumentService: TextDocumentService = new RiddlLSPTextDocumentService()
+  override def getTextDocumentService: TextDocumentService = documentService
 
-  override def getWorkspaceService: WorkspaceService = new RiddlLSPWorkspaceService()
+  override def getWorkspaceService: WorkspaceService = workspaceService
 }
