+++ server/src/main/scala/com/ossuminc/riddl/lsp/server/RiddlLSPTextDocumentService/Vars.scala
@@ -3,7 +3,7 @@ package com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService
 import com.ossuminc.riddl.language.AST
 import com.ossuminc.riddl.language.Messages.Messages
 
-object vars {
+object Vars {
   var docLines: Seq[String] = Seq()
   var docAST: Option[Either[Messages, AST.Root]] = None
   var docURI: Option[String] = None
