+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocContentMgmtSpec/ActionsAndLensesSpec.scala
@@ -1,15 +1,20 @@
 package com.ossum.riddl.lsp.server.LSPTextDocumentSpec.DocContentMgmtSpec
 
-import com.ossum.riddl.lsp.server.initializationSpecs.{
+import com.ossum.riddl.lsp
+import com.ossum.riddl.lsp.server.InitializationSpecs.{
   OpenNoErrorFileSpec,
   OpenOneErrorFileSpec
 }
-import com.ossum.riddl.lsp.server.requestSpecs.DiagnosticRequestSpec
+import com.ossum.riddl.lsp.server.RequestSpecs.DiagnosticRequestSpec
+import com.ossuminc.riddl.lsp.server.RiddlLSPServer
 import org.eclipse.lsp4j.DocumentDiagnosticReportKind
+import org.scalatest.ParallelTestExecution
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
+import java.io.File
+import java.nio.file.Files
 import scala.jdk.FutureConverters.*
 import scala.jdk.CollectionConverters.*
 
@@ -19,10 +24,16 @@ import scala.jdk.CollectionConverters.*
 - codeResolve
  */
 
-class ActionsAndLensesSpec extends AnyWordSpec with Matchers with ScalaFutures {
+class ActionsAndLensesSpec
+    extends AnyWordSpec
+    with Matchers
+    with ScalaFutures
+    with ParallelTestExecution {
 
   "fail requesting for diagnostic from file with no errors" in new OpenNoErrorFileSpec
     with DiagnosticRequestSpec {
+
+    Files.delete(tempFilePath)
 
     diagnosticResultF.asScala.failed.futureValue mustBe a[Throwable]
     diagnosticResultF.asScala.failed.futureValue.getMessage mustEqual "Document has no errors"
@@ -30,6 +41,8 @@ class ActionsAndLensesSpec extends AnyWordSpec with Matchers with ScalaFutures {
 
   "succeed requesting for diagnostic from file with one error" in new OpenOneErrorFileSpec
     with DiagnosticRequestSpec {
+
+    Files.delete(tempFilePath)
 
     whenReady(diagnosticResultF.asScala) { report =>
       report.getRelatedFullDocumentDiagnosticReport.getKind mustEqual DocumentDiagnosticReportKind.Full
