+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocContentMgmtSpec/FormattingAndRenamingSpec.scala
@@ -1,5 +1,7 @@
 package com.ossum.riddl.lsp.server.LSPTextDocumentSpec.DocContentMgmtSpec
 
+import com.ossum.riddl.lsp.server
+import org.scalatest.ParallelTestExecution
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpec
@@ -12,6 +14,8 @@ import org.scalatest.wordspec.AnyWordSpec
 - prepareRename
  */
 
-class FormattingAndRenamingSpec extends AnyWordSpec with Matchers with ScalaFutures {
-
-}
+class FormattingAndRenamingSpec
+    extends AnyWordSpec
+    with Matchers
+    with ScalaFutures
+    with ParallelTestExecution {}
