+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocContentMgmtSpec/LinksAndColoursSpec.scala
@@ -1,5 +1,7 @@
 package com.ossum.riddl.lsp.server.LSPTextDocumentSpec.DocContentMgmtSpec
 
+import com.ossum.riddl.lsp
+import org.scalatest.ParallelTestExecution
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpec
@@ -11,6 +13,8 @@ import org.scalatest.wordspec.AnyWordSpec
 - colorPresentation
  */
 
-object LinksAndColoursSpec extends AnyWordSpec with Matchers with ScalaFutures {
-  
-}
+object LinksAndColoursSpec
+    extends AnyWordSpec
+    with Matchers
+    with ScalaFutures
+    with ParallelTestExecution {}
