+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/DocContentMgmtSpec/SymbolInfoSpec.scala
@@ -1,5 +1,7 @@
 package com.ossum.riddl.lsp.server.LSPTextDocumentSpec.DocContentMgmtSpec
 
+import com.ossum.riddl.lsp
+import org.scalatest.ParallelTestExecution
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpec
@@ -14,6 +16,8 @@ import org.scalatest.wordspec.AnyWordSpec
 - documentSymbol
  */
 
-object SymbolInfoSpec extends AnyWordSpec with Matchers with ScalaFutures {
-  
-}
+object SymbolInfoSpec
+    extends AnyWordSpec
+    with Matchers
+    with ScalaFutures
+    with ParallelTestExecution {}
