+++ server/src/test/scala/com/ossum/riddl/lsp/server/LSPTextDocumentSpec/WorkspaceMgmtSpec.scala
@@ -1,7 +1,12 @@
 package com.ossum.riddl.lsp.server.LSPTextDocumentSpec
 
+import org.scalatest.ParallelTestExecution
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
-class WorkspaceMgmtSpec extends AnyWordSpec with Matchers with ScalaFutures {}
+class WorkspaceMgmtSpec
+    extends AnyWordSpec
+    with Matchers
+    with ScalaFutures
+    with ParallelTestExecution {}
