+++ server/src/test/scala/com/ossum/riddl/lsp/server/RiddlLSPServerSpec.scala
@@ -2,14 +2,18 @@ package com.ossum.riddl.lsp.server
 
 import com.ossuminc.riddl.lsp.server.RiddlLSPTextDocumentService.RiddlLSPTextDocumentService
 import com.ossuminc.riddl.lsp.server.{RiddlLSPServer, RiddlLSPWorkspaceService}
-import com.ossuminc.riddl.lsp.utils.createInitializeResultIncremental
+import com.ossuminc.riddl.lsp.utils.Utils.createInitializeResultIncremental
 import org.eclipse.lsp4j.{InitializeParams, InitializeResult}
+import org.scalatest.ParallelTestExecution
 import org.scalatest.matchers.must.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 
 import scala.language.postfixOps
 
-class RiddlLSPServerSpec extends AnyWordSpec with Matchers {
+class RiddlLSPServerSpec
+    extends AnyWordSpec
+    with Matchers
+    with ParallelTestExecution {
 
   trait InitializeSpec {
     val server: RiddlLSPServer = RiddlLSPServer()
@@ -32,11 +36,11 @@ class RiddlLSPServerSpec extends AnyWordSpec with Matchers {
     }
 
     "getTextDocumentService properly" in new InitializeSpec {
-      server.getTextDocumentService() mustBe a [RiddlLSPTextDocumentService]
+      server.getTextDocumentService() mustBe a[RiddlLSPTextDocumentService]
     }
 
     "getWorkspace properly" in new InitializeSpec {
-      server.getWorkspaceService() mustBe a [RiddlLSPWorkspaceService]
+      server.getWorkspaceService() mustBe a[RiddlLSPWorkspaceService]
     }
   }
